raftNode: add a new and better way to tell when the node is done
The old method, monitoring the `errorC` channel, is not great because
the error only pops out of that channel once. Which of the pieces of
code that are waiting on that channel reads the error? Nobody knows!

Instead, provide a `Done()` method and an `Err()` method that
interested parties can use to determine when the node finishes and
what error it returned.

The callers haven't been changed yet.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>