Allow applying entries on durable remote quorum
Currently, a node will only emit entries for application that are
durably in the log locally.

To provide high availability in the presence of grey failures such as
disk stalls, it is desirable to be able to apply entries that are not
durable locally but are nevertheless known to have been persisted on a
quorum.

This PR prototypes such a mechanism which is opt-in and only available
in conjunction with with async storage writes.

Applying entries that aren't in the local log requires some care:

- if snapshots are sent based on the applied state, the leader may have
  to catch up the follower from its unstable storage. This is fine. If
the leader crashes, the next leader will definitely have a durable log
that will catch up the snapshot recipient thanks to the log completeness
property. (i.e. the next leader is from the quorum).
- orchestrating log truncation via a side effect of applying an entry is
  more complicated since the truncation may now be carried out on a
portion of the log that will only be written in the future.

I haven't thought much about other problems that may occur, since at
present I'm sending this mostly for prototyping and discussion, not
with an intention to merge.