add `RawNode.CampaignNow()`
This patch adds `RawNode.CampaignNow()`, which can be used to bypass the
CheckQuorum condition that votes are only granted if the voter hasn't
heard from the leader in the past election timeout interval. PreVote is
otherwise applied and respected as usual, i.e. prevotes are not granted
if the candidate is behind on its log.

In CockroachDB, this is useful when we know the leader no longer exists
(e.g. because it was removed from the Raft group), or if we need to
fast-track an election for a quiesced, inactive range that hasn't been
ticked in a while.

Signed-off-by: Erik Grinaker <grinaker@cockroachlabs.com>